# backend/docs/openapi.v0.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "duoduo-product-testing-api",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/wechat/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Wechat Login",
        "description": "Mock WeChat mini-program login.",
        "operationId": "wechat_login_api_v1_auth_wechat_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WechatLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/profile": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update Profile",
        "description": "Update the authenticated user's profile.",
        "operationId": "update_profile_api_v1_auth_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/avatar": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Upload Avatar",
        "description": "Upload and set the authenticated user's avatar image.",
        "operationId": "upload_avatar_api_v1_auth_avatar_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_auth_avatar_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Token",
        "description": "Refresh the authenticated user's token.",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Logout the current user by revoking the active bearer token.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Logout Api V1 Auth Logout Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "description": "Return the authenticated user.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/credits/balance": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "Get Balance",
        "description": "Return the authenticated user's current credit balance.",
        "operationId": "get_balance_api_v1_credits_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditBalanceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/credits/transactions": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "List Transactions",
        "description": "Return the authenticated user's credit transaction history, newest-first.",
        "operationId": "list_transactions_api_v1_credits_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditTransactionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvatarUploadResponse": {
        "properties": {
          "avatar_url": {
            "type": "string",
            "title": "Avatar Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "avatar_url"
        ],
        "title": "AvatarUploadResponse",
        "description": "Avatar upload response."
      },
      "Body_upload_avatar_api_v1_auth_avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_avatar_api_v1_auth_avatar_post"
      },
...
      "CreditBalanceResponse": {
        "properties": {
          "balance": {
            "type": "integer",
            "title": "Balance"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "balance",
          "updated_at"
        ],
        "title": "CreditBalanceResponse",
        "description": "Response for GET /credits/balance."
      },
      "CreditTransactionItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "balance_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance After"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "ref_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Type"
          },
          "ref_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "amount",
          "balance_after",
          "reason",
          "ref_type",
          "ref_id",
          "note",
          "created_at"
        ],
        "title": "CreditTransactionItem",
        "description": "One credit transaction row."
      },
      "CreditTransactionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CreditTransactionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "CreditTransactionListResponse",
        "description": "Cursor-paginated credit transaction list."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginUserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "credit_balance": {
            "type": "integer",
            "title": "Credit Balance"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "nickname",
          "avatar_url",
          "role_type",
          "credit_balance",
          "is_new_user"
        ],
        "title": "LoginUserResponse",
        "description": "User response shape for login."
      },
...
      "ProfileUpdateRequest": {
        "properties": {
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileUpdateRequest",
        "description": "Profile update request."
      },
...
      "RefreshTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "expires_in"
        ],
        "title": "RefreshTokenResponse",
        "description": "JWT refresh response."
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "credit_balance": {
            "type": "integer",
            "title": "Credit Balance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "nickname",
          "avatar_url",
          "role_type",
          "credit_balance"
        ],
        "title": "UserResponse",
        "description": "User response shape defined by the API contract."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WechatLoginRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 128,
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "WechatLoginRequest",
        "description": "Wechat mini-program login request."
      },
      "WechatLoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/LoginUserResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "expires_in",
          "user"
        ],
        "title": "WechatLoginResponse",
        "description": "Wechat login response."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}